requirements.txt
rembg~=2.0.25
Pillow~=9.2.0
#aiogram~=2.22.2